# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PKL Attendance API",
        "description": "API untuk sistem absensi PKL menggunakan WhatsApp Bot & Web Admin",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api-production.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/attendances": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get attendance list with filters",
                "operationId": "57cd91ddd513b3298d731adc13ba7560",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by date (Y-m-d format), defaults to today",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-02-02"
                        }
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "Filter by member ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by attendance status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "hadir",
                                "izin",
                                "sakit",
                                "alpha"
                            ]
                        }
                    },
                    {
                        "name": "office_id",
                        "in": "query",
                        "description": "Filter by office ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attendances/{id}": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get specific attendance record",
                "operationId": "a71023f5c8c3443d5396f2dd0af2d0dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Attendance ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Attendance not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attendances/check-in": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Check-in endpoint (public)",
                "description": "Used by WhatsApp bot or manual check-in. Validates geofencing.",
                "operationId": "cd659c287d106562fad023b8ad183d1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "no_hp",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "no_hp": {
                                        "type": "string",
                                        "example": "+6281234567890"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.20005
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.8167
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Check-in successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Check-in berhasil!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already checked in"
                    },
                    "403": {
                        "description": "Outside geofence area"
                    },
                    "404": {
                        "description": "Member not found or inactive"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/attendances/check-out": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Check-out endpoint (public)",
                "description": "Used by WhatsApp bot or manual check-out. Validates geofencing and calculates working hours.",
                "operationId": "3cdb7b994d68c55886732da1b8f8bb74",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "no_hp",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "no_hp": {
                                        "type": "string",
                                        "example": "+6281234567890"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.20005
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.8167
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-out successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Check-out berhasil!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not checked in yet or already checked out"
                    },
                    "403": {
                        "description": "Outside geofence area"
                    },
                    "404": {
                        "description": "Member not found or inactive"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token dari endpoint /api/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}